requirements.txt
requests~=2.32.3
psutil~=6.0.0
pystray~=0.19.5
pillow~=10.4.0
pypresence~=4.2.0
# Core dependencies
psutil>=5.9.0
pillow>=9.0.0
pystray>=0.19.0
requests>=2.27.0
configparser>=5.2.0
pypresence>=4.2.0

# Build dependencies
pyinstaller>=5.6.0
